AMQNET-585: NMS AMQP .NetStd 2.0 Client Support
- Add .NET Standard support
- Add nuget package configuration
- Update dependencies
- Add Microsoft.NET.Test.Sdk to be able to run tests directly form VS
- Get the directory containing the current test assembly properly
- Fix race condition on .NET Core

In previous implementation, CountDownLatch was instantiated after
OpenResponse subscription was made. It might lead to race conditions
when the connection was established before proper instance CountDownLatch
was set. As a result attempt to call Connect would end up with timeout.
For whatever reasons it wasn't a case with .NET Framework set as
a target framework, but happens every single time when you try to run
tests against .NET Core runtime.